+++ api/include/mk_tcp.h
@@ -1,7 +1,7 @@
 ﻿/*
  * Copyright (c) 2016 The ZLMediaKit project authors. All Rights Reserved.
  *
- * This file is part of ZLMediaKit(https://github.com/xiongziliang/ZLMediaKit).
+ * This file is part of ZLMediaKit(https://github.com/xia-chu/ZLMediaKit).
  *
  * Use of this source code is governed by MIT license that can be found in the
  * LICENSE file in the root of the source tree. All contributing project authors
@@ -53,9 +53,9 @@ API_EXPORT mk_sock_info API_CALL mk_tcp_session_get_sock_info(const mk_tcp_sessi
 //TcpSession::safeShutdown()
 API_EXPORT void API_CALL mk_tcp_session_shutdown(const mk_tcp_session ctx,int err,const char *err_msg);
 //TcpSession::send()
-API_EXPORT void API_CALL mk_tcp_session_send(const mk_tcp_session ctx,const char *data,int len);
+API_EXPORT void API_CALL mk_tcp_session_send(const mk_tcp_session ctx,const char *data,size_t len);
 //切换到该对象所在线程后再TcpSession::send()
-API_EXPORT void API_CALL mk_tcp_session_send_safe(const mk_tcp_session ctx,const char *data,int len);
+API_EXPORT void API_CALL mk_tcp_session_send_safe(const mk_tcp_session ctx,const char *data,size_t len);
 
 ///////////////////////////////////////////自定义tcp服务/////////////////////////////////////////////
 
@@ -74,7 +74,7 @@ typedef struct {
      * @param data 数据指针
      * @param len 数据长度
      */
-    void (API_CALL *on_mk_tcp_session_data)(uint16_t server_port,mk_tcp_session session,const char *data,int len);
+    void (API_CALL *on_mk_tcp_session_data)(uint16_t server_port,mk_tcp_session session,const char *data,size_t len);
 
     /**
      * 每隔2秒的定时器，用于管理超时等任务
@@ -164,7 +164,7 @@ typedef struct {
      * @param data 数据指针
      * @param len 数据长度
      */
-    void (API_CALL *on_mk_tcp_client_data)(mk_tcp_client client,const char *data,int len);
+    void (API_CALL *on_mk_tcp_client_data)(mk_tcp_client client,const char *data,size_t len);
 
     /**
      * 每隔2秒的定时器，用于管理超时等任务
