+++ src/Common/Parser.h
@@ -1,7 +1,7 @@
 ﻿/*
  * Copyright (c) 2016 The ZLMediaKit project authors. All Rights Reserved.
  *
- * This file is part of ZLMediaKit(https://github.com/xiongziliang/ZLMediaKit).
+ * This file is part of ZLMediaKit(https://github.com/xia-chu/ZLMediaKit).
  *
  * Use of this source code is governed by MIT license that can be found in the
  * LICENSE file in the root of the source tree. All contributing project authors
@@ -19,7 +19,7 @@ using namespace toolkit;
 
 namespace mediakit{
 
-string FindField(const char *buf, const char *start, const char *end, int bufSize = 0);
+string FindField(const char *buf, const char *start, const char *end, size_t bufSize = 0);
 
 struct StrCaseCompare {
     bool operator()(const string &__x, const string &__y) const {
