+++ src/Rtmp/RtmpDemuxer.cpp
@@ -1,7 +1,7 @@
 ﻿/*
  * Copyright (c) 2016 The ZLMediaKit project authors. All Rights Reserved.
  *
- * This file is part of ZLMediaKit(https://github.com/xiongziliang/ZLMediaKit).
+ * This file is part of ZLMediaKit(https://github.com/xia-chu/ZLMediaKit).
  *
  * Use of this source code is governed by MIT license that can be found in the
  * LICENSE file in the root of the source tree. All contributing project authors
@@ -25,7 +25,7 @@ bool RtmpDemuxer::loadMetaData(const AMFValue &val){
         const AMFValue *videocodecid = nullptr;
         val.object_for_each([&](const string &key, const AMFValue &val) {
             if (key == "duration") {
-                _fDuration = val.as_number();
+                _fDuration = (float)val.as_number();
                 return;
             }
             if (key == "audiosamplerate") {
