+++ src/Rtp/GB28181Process.cpp
@@ -1,7 +1,7 @@
 ﻿/*
  * Copyright (c) 2016 The ZLMediaKit project authors. All Rights Reserved.
  *
- * This file is part of ZLMediaKit(https://github.com/xiongziliang/ZLMediaKit).
+ * This file is part of ZLMediaKit(https://github.com/xia-chu/ZLMediaKit).
  *
  * Use of this source code is governed by MIT license that can be found in the
  * LICENSE file in the root of the source tree. All contributing project authors
@@ -18,7 +18,7 @@
 namespace mediakit{
 
 //判断是否为ts负载
-static inline bool checkTS(const uint8_t *packet, int bytes){
+static inline bool checkTS(const uint8_t *packet, size_t bytes){
     return bytes % TS_PACKET_SIZE == 0 && packet[0] == TS_SYNC_BYTE;
 }
 
@@ -30,7 +30,7 @@ GB28181Process::GB28181Process(const MediaInfo &media_info, MediaSinkInterface *
 
 GB28181Process::~GB28181Process() {}
 
-bool GB28181Process::inputRtp(bool, const char *data, int data_len) {
+bool GB28181Process::inputRtp(bool, const char *data, size_t data_len) {
     return handleOneRtp(0, TrackVideo, 90000, (unsigned char *) data, data_len);
 }
 
@@ -73,7 +73,7 @@ void GB28181Process::onRtpSorted(const RtpPacket::Ptr &rtp, int) {
     _rtp_decoder->inputRtp(rtp, false);
 }
 
-const char *GB28181Process::onSearchPacketTail(const char *packet,uint64_t bytes){
+const char *GB28181Process::onSearchPacketTail(const char *packet,size_t bytes){
     try {
         auto ret = _decoder->input((uint8_t *) packet, bytes);
         if (ret > 0) {
