+++ src/Rtp/PSDecoder.cpp
@@ -1,7 +1,7 @@
 ﻿/*
  * Copyright (c) 2016 The ZLMediaKit project authors. All Rights Reserved.
  *
- * This file is part of ZLMediaKit(https://github.com/xiongziliang/ZLMediaKit).
+ * This file is part of ZLMediaKit(https://github.com/xia-chu/ZLMediaKit).
  *
  * Use of this source code is governed by MIT license that can be found in the
  * LICENSE file in the root of the source tree. All contributing project authors
@@ -44,7 +44,7 @@ PSDecoder::~PSDecoder() {
     ps_demuxer_destroy((struct ps_demuxer_t*)_ps_demuxer);
 }
 
-int PSDecoder::input(const uint8_t *data, int bytes) {
+size_t PSDecoder::input(const uint8_t *data, size_t bytes) {
     return ps_demuxer_input((struct ps_demuxer_t*)_ps_demuxer,data,bytes);
 }
